requirements.txt
matplotlib==3.4.3
numpy==1.20.3
opencv_python==4.7.0.68
Pillow==9.4.0
tensorflow==2.11.0
typer==0.7.0
